15/15 on Zybooks
The comments for DFS and DFSUtil might be reversed, confusion on how to separate the algorithm because in this lab we don't recursively call Depth First Search on an entire graph, we are only calling it once expecting the graph to be fully connected. If the graph is not fully traversable with a single call then we know the graph isn't connected. Calling both functions would result in a "forrest", if the graph is not connected, and eventually would visit all vertices in the graph turning the bool values in mvistedVertex true. This isn't what we want for this lab.